requirements.txt
flask==2.2.5
cloudinary==1.40.0
python_dotenv==0.21.1